# openapi.json
{"openapi":"3.1.0","info":{"title":"Kaguya Music Mood API","description":"AI-powered mood detection from video stream with Spotify playlist recommendations","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"Root endpoint with API information","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/detect-mood":{"post":{"summary":"Detect Mood","description":"Detect mood from base64 encoded image","operationId":"detect_mood_detect_mood_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoodDetectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoodDetectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/playlist/{mood}":{"get":{"summary":"Get Playlist By Mood","description":"Get Spotify playlist recommendations based on mood","operationId":"get_playlist_by_mood_playlist__mood__get","parameters":[{"name":"mood","in":"path","required":true,"schema":{"type":"string","title":"Mood"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaylistResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mood-and-playlist":{"post":{"summary":"Detect Mood And Get Playlist","description":"Detect mood from image and return Spotify playlist recommendations","operationId":"detect_mood_and_get_playlist_mood_and_playlist_post","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoodDetectionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoodDetectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/moods":{"get":{"summary":"Get Available Moods","description":"Get list of available moods","operationId":"get_available_moods_moods_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/upload-image":{"post":{"summary":"Upload Image Mood Detection","description":"Upload image file for mood detection (alternative to base64)","operationId":"upload_image_mood_detection_upload_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_image_mood_detection_upload_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_image_mood_detection_upload_image_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_image_mood_detection_upload_image_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MoodDetectionRequest":{"properties":{"image_base64":{"type":"string","title":"Image Base64"}},"type":"object","required":["image_base64"],"title":"MoodDetectionRequest"},"MoodDetectionResponse":{"properties":{"mood":{"type":"string","title":"Mood"},"confidence":{"type":"number","title":"Confidence"},"playlist_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Playlist Url"},"recommendations":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Recommendations","default":[]}},"type":"object","required":["mood","confidence"],"title":"MoodDetectionResponse"},"PlaylistResponse":{"properties":{"mood":{"type":"string","title":"Mood"},"playlist_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Playlist Url"},"tracks":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Tracks","default":[]}},"type":"object","required":["mood"],"title":"PlaylistResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}